requirements.txt
# ============================================================
# ForgeForth Africa - Django Platform Requirements
# Compatible with Python 3.9+
# Using Django 4.2 LTS (supports Python 3.8 - 3.12)
# ============================================================

# Core
Django>=4.2,<4.3
django-environ==0.11.2

# REST API
djangorestframework==3.14.0
drf-spectacular==0.27.1
djangorestframework-simplejwt==5.3.1

# CORS
django-cors-headers==4.3.1

# Database drivers
# ─────────────────────────────────────────────────────────────────────────────
# cPanel / shared hosting (MySQL/MariaDB — provided natively by cPanel):
mysqlclient>=2.2.0          # MySQL adapter — install on server via pip
PyMySQL>=1.1.0              # Pure-Python fallback if mysqlclient can't compile
# PostgreSQL for Neon cloud sync:
psycopg2-binary>=2.9.9      # PostgreSQL adapter for Neon sync

# Caching
django-redis==5.4.0

# WebSockets / Channels (WSS support)
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Celery - Background Tasks
celery==5.3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.1

# HTTP Client with SSL support
httpx==0.27.0

# Email
django-anymail==10.2

# Storage
django-storages==1.14.2
boto3==1.34.34

# Auth & Security
django-axes==6.3.0
django-allauth==0.61.1
PyJWT==2.8.0
cryptography==42.0.2

# Admin
django-unfold==0.22.0
django-import-export==3.3.6

# Audit
django-auditlog==2.3.0

# Filtering & Search
django-filter==23.5
elasticsearch-dsl==8.11.0

# Utilities
django-extensions==3.2.3
python-dotenv==1.0.0
python-slugify==8.0.4
phonenumbers==8.13.30
pycountry==23.12.11

# File Processing & Media
python-magic==0.4.27
python-docx==1.1.0

# Media Processing Pipelines
# ─────────────────────────────────────────────────────────────────────────────
# Images: sanitization, compression, WebP conversion
Pillow>=10.2.0
pillow-avif-plugin>=1.4.0

# Documents: PDF text extraction, thumbnails
PyMuPDF>=1.23.0           # fitz library for PDF processing

# Audio: compression, metadata stripping
mutagen>=1.47.0           # Audio metadata handling
pydub>=0.25.1             # Audio processing (needs ffmpeg on server)

# Video: transcoding, thumbnails (needs ffmpeg on server)
ffmpeg-python>=0.2.0      # Python wrapper for ffmpeg

# HTTP Client
httpx==0.27.0
requests==2.31.0

# Monitoring & Logging
sentry-sdk==1.40.5
django-structlog==6.0.1

# Jinja2 (used for website templates)
Jinja2>=3.1

# Translation Services
googletrans==4.0.0-rc1  # Free Google Translate (rate limited)
beautifulsoup4==4.12.3  # HTML parsing for translation
lxml==5.1.0             # Fast HTML/XML parser
# Optional: deepl==1.17.0  # DeepL API (requires paid key)
# Optional: google-cloud-translate==3.15.0  # Google Cloud Translate (paid)

# Static Files
whitenoise==6.6.0

# Production
gunicorn==21.2.0

# ============================================================
# Auth Microservice (FastAPI-based)
# ============================================================
fastapi==0.109.2
uvicorn[standard]==0.27.0
pydantic[email]==2.6.0
python-multipart==0.0.9
pyotp==2.9.0
werkzeug==3.0.1
SQLAlchemy==2.0.25
